Dados são perdidos ao voltar, cancelar ou quando é apresentada uma modal com questionamento AB#10125
